refactor: extract physics and statistics from order_analysis
Extract rotational-frequency calculators + OrderHypothesis catalog into
rotational_physics.py (~156 lines) and statistical evidence functions into
order_statistics.py (~259 lines). order_analysis.py shrinks from 1167 to
~819 lines and now focuses on matching, heuristic filters, and assembly.

Closes #710
